feat: Add getcountasync to repository and fix totalpages math

+++ CobreezeHousehold/Application/Wrappers/PagedResponse.cs
@@ -4,13 +4,10 @@ namespace Application.Wrappers
 {
     public class PagedResponse<T> : Response<T>
     {
-        /*
-         * TODO: totalrecords is the count of the already filtered collection,..
-         */
         public int PageNumber { get; set; }
         public int PageSize { get; set; }
         public int TotalRecords { get; set; }
-        public int TotalPages => (int)Math.Ceiling((double)(TotalRecords / PageSize));
+        public int TotalPages => (int)Math.Ceiling((double)TotalRecords / (double)PageSize);
 
         public PagedResponse(T data, int pageNumber, int pageSize, int totalRecords) : base(data)
         {
